Added patches for the "Magic Fang" issue. KM generally handles polymorph weapon slots and additional limbs poorly, and they are missed from PreSave/PostSave/Dispose/TurnOn/TurnOff events. By properly turning them on, their enchantments get applied, and magic fang works after save/load.
Could not fix the Feral Mutagen issue, because it's actually a different problem. The Feral Mutagen buff has an EmptyHandWeaponOverride component which sets the Empty Hand Weapon on TurnOn/TurnOff. The problem is that it's a new weapon instance each time, and buffs (facts) are not preserved. WotR fixes this by having serialized data in action components, which is not something we have in KM. Putting this off for now