Fix Feral Mutagen losing claw enchantments on save. The EmptyHandWeaponOverride game logic component caches its empty hand weapon during program lifetime, but previously was unable to reapply it when being turned off and back on again during save. Now, the existing claw is restored if still cached.
For the load issue, we'd need the empty hand weapon to be serialized in the save, which would require game logic components to have any save data. WotR does implement such save data, but I'm not sure it could be done in KM with Harmony patches. Out of scope for now